Move accessor related types to accessor.hh.
